create new select and params note

+++ src/views/Home/components/common/NoteParamsOverlay.vue
@@ -23,13 +23,12 @@
                         max-w-2xl max-h-140 lg:max-h-100 
                         lg:grid lg:grid-cols-[280px_1em_300px]
                         flex justify-center items-center
-                        w-full h-full gap-10 
+                        w-full gap-10 
                         pointer-events-auto
-                        translate-y-[-20%]
                     "
                 >
 
-                    <div class="w-full h-full hidden lg:block">
+                    <div class="w-full hidden lg:block">
 
                         <DefaultNoteCard
                             v-if="note && showCard"
@@ -117,8 +116,8 @@
                                                     :id="`switch-${tag.id}`"
                                                     type="checkbox"
                                                     class="sr-only peer"
-                                                    :checked="note?.tags.includes(tag.id)"
-                                                    @change="toggleTag(tag.id)"
+                                                    :checked="justView ? note?.tags?.includes(tag.id) : tags.includes(tag.id)"
+                                                    @change="justView ? () => {} : toggleTag(tag.id)"
                                                 />
 
                                                 <div
@@ -238,7 +237,10 @@
                             <hr class="-mx-4 text-gray-300" />
 
                             <div class="">
-                                <button class="w-full primary" @click="save">Sauvegarder</button>
+                                <button 
+                                    class="w-full primary"
+                                    @click="save"
+                                >Sauvegarder</button>
                             </div>
 
                         </div>
@@ -252,6 +254,7 @@
         </transition>
 
         <Share_menu
+            v-if="uuid"
             :uuid="uuid"
             v-model="showShareMenu"
         />
@@ -272,7 +275,7 @@
 
 import BackdropOverlay from '@/components/common/BackdropOverlay.vue';
 import type { Note } from '@/assets/ts/type';
-import { nextTick, ref, watch } from 'vue';
+import { nextTick, onMounted, ref, watch } from 'vue';
 import { Notes, Tags } from '@/assets/ts/database/Var';
 import DefaultNoteCard from '@/views/Home/components/common/DefaultNoteCard.vue';
 import database from '@/assets/ts/database/database';
@@ -282,16 +285,22 @@ import ConfirmDialog from '@/components/popup/ConfirmDialog.vue';
 
 const props = defineProps<{
     visible: Boolean;
-    uuid: string;
+    uuid?: string;
     justTags?: boolean;
+    justView?: boolean;
+    selectedTags?: number[];
 }>();
 
-const note = ref<Note | undefined>(undefined);
+const note = ref<Note | undefined | any>(undefined);
 const showCard = ref<boolean>(false);
 const showShareMenu = ref<boolean>(false);
 const showConfirmDel = ref<boolean>(false);
+const tags = ref<number[]>(props.selectedTags || []);
 
-const emit = defineEmits(['update:visible']);
+const emit = defineEmits([
+    'update:visible',
+    'TagsCallback'
+]);
 
 const emitClose = () => {
   emit('update:visible', false);
@@ -306,6 +315,7 @@ const deleteNote = async (state: number) => {
     else
     {
         showConfirmDel.value = false;
+        if (!props.uuid) return;
         await database.delete(props.uuid);
         Notes.value = Notes.value.filter(_note => _note.uuid !== props.uuid);
         emitClose();
@@ -313,7 +323,14 @@ const deleteNote = async (state: number) => {
 }
 
 const save = () => {
+    
     if (!note.value) return;
+
+    if (props.justView) {
+        emit('TagsCallback', tags.value);
+        emit('update:visible', false);
+        return;
+    }
 
     const index = Notes.value.findIndex(n => n.uuid === props.uuid);
 
@@ -362,15 +379,26 @@ const refrechCard = async () => {
 
 
 const mount = () => {
+    
+    if (props.justView) {
+        note.value = {};
+        tags.value = props.selectedTags || [];
+        showCard.value = true;
+        return;
+    }
+
     const _note = Notes.value.find(note => note.uuid === props.uuid);
     note.value = JSON.parse(JSON.stringify(_note));
     showCard.value = true;
     if (!note) throw new Error('Note is undefined on params overlay !')
+
 }
 
 watch(() => props.visible, () => {
     if (props.visible) mount()
 })
+
+onMounted(mount);
 
 
 </script>
